Load capabilities when nexus is initialized instead of when nexus is started
Add UT for nexus initialized event inspector.
